Refactoring of the `graphql-api` service to use ports for `Database` (#858)
Refs https://github.com/FuelLabs/fuel-core/issues/809
Closes: https://github.com/FuelLabs/fuel-core/issues/907

It is the refactoring for the `graphql-api` service to use ports instead
of direct work with `Database`. The service defines minimal database
functionality and builds its functionality on top of it to process
GrapQL requests.

Right now, this functionality is represented by `{}QueryContext`
structures with the public methods. It is done to make the migration
easier.
But later, we will need to move this functionality into corresponding
traits and implement it directly for the `DatabaseTemp` type. Like:

```rust
impl {}QueryContext for DatabaseTemp {...}

or

impl<D: DatabasePort> {}QueryContext for D {...}
```

With traits, we can mock them and improve our tests.
The PR contains the following changes:

- Renamed `Transactional` into `Transaction`.
- Added `Transactional` trait with `transaction` method.
- Used `BlockId` instead of `Bytes32` in `Database`. 
- Created ports for `block.rs`, `tx.rs`, `message.rs`, `balance.rs`,
`coin.rs` and updated corresponding schema.
- Added and used in ports `BoxedIter` instead of associated types(In
most cases they were `Box<dyn>` too).
- Almost removed usage of fuel_core::database from schema. Only `dap.rs`
uses it right now for `VMDatabase`.
- Introduced `CompressedCoin` and `Coin`(contains `UtxoId`).
`CompressedCoin` is used during work with the database, `Coin` in
external API.
- Moved resources into `fuel-core-type` and simplified it by removing
generics.
- Tried to re-work `dap.rs` but the change is big and also related to
`fuel-core-executor`, so will do it in a separate PR.
- Updated `Transactional` to support `Transactional<dyn Database>`

Before we can extract `grapgql-api` into its crate, we need:
- Add ports for other services.
- Add iteration abstraction into `fuel-core-storage`.
- Solve the problem with `DatabaseVM`(The solution should also fit for
`fuel-core-executor`).

Co-authored-by: Brandon Kite <brandonkite92@gmail.com>
Co-authored-by: Green Baneling <XgreenX9999@gmail.com>