Connect the block importer with outside world (#915)
- Closes https://github.com/FuelLabs/fuel-core/issues/887
- Closes https://github.com/FuelLabs/fuel-core/issues/884

What is done:

- Removed `CHAIN_HEIGHT_KEY`, because now we have a genesis block to
track this information.
- Unified all database API that works with `BlockHeight` to work with
references. Previously in some places, it was the reference, in some
not.
- Added `manual_produce_block` into `PoA` and used it in the `GraphQL`
to have the one source of the block production. There are no unit
tests(help is appreciated).
- Connected `PoA` and `BlockImporter` via ports. Added
`BlockImporterAdapter`.
- Added the `FuelBlockRoots` column with the corresponding table. It is
used to track BMT MMR roots for fuel blocks. The implementation adds new
`DatabaseColumn` and `IntoDatabaseKey` traits to automate the storage
trait implementation for some tables.
- Moved implementation of database ports into the corresponding file in
the `fuel_core::service::adapter`.
- Removed the `Database` generic from `PoA` at all. Now we must pass the
`last_height` while creating the `Task`. Updated tests to work with it.
- Block producer now accepts the `block_time` along with `block_height`.

Co-authored-by: Tom <tomrgowan@gmail.com>
Co-authored-by: ControlCplusControlV <44706811+ControlCplusControlV@users.noreply.github.com>
Co-authored-by: Voxelot <brandonkite92@gmail.com>