Fix database version serialization (#1015)
The database version was being double initialized between
`Database::init` and `RocksDB`. However, the `Database` was using
postcard, and `RocksDB` was using manual byte serialization. This was
causing version conflicts when restarting a fuel-core node off of an
existing database, as the serialization methods were mismatched.

This PR resolves the issue by ensuring the database metadata is only
managed by the database itself, and not by RocksDb.

This is a blocker for beta3.