feat: Add `enable-p2p` flag to `fuel-core` CLI (#1268)
Related issues:
- https://github.com/FuelLabs/fuel-core/issues/1236

This PR disables P2P by default, and adds a new flag to the `fuel-core`
CLI to allow users to expressly enable P2P:
```
--enable-p2p
```
Currently, the P2P service is only enabled when the user provides CLI
arguments for `keypair` and `network`. If either of these values are
missing, P2P is implicitly disabled without warning to the user. There
is no way to tell if the user accidentally omitted these values, or
deliberately omitted them in order to disable P2P.

A better user experience is to enable P2P explicitly using
`--enable-p2p` and to make `keypair` and `network` mandatory. This will
raise an error if either of these values is missing. P2P should be
disabled by default. This flag replaces having to add or remove the
`network` argument to enable or disable P2P implicitly.

If a user wants to run `fuel-core` with P2P, and supplies
`--enable-p2p`, but forgets to include `keypair` and/or `network`, they
will encounter an error:
```
error: the following required arguments were not provided:
  --keypair <KEYPAIR>
  --network <NETWORK>
  ```

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>