Block gas limit calibration (#1273)
- Added a new benchmark, `block_target_gas`, that runs different simple
scripts with opcode and measures the block execution time. It allows us
to track how accurate is our gas cost. If each benchmark takes ~ the
same amount of time, then our gas cost is accurate. The current costs is
not accurate because the `noop` takes ~1s for 100M gas, while `meq` -
~6s, `logd` - ~23s... Later we can have more opcodes tested and the end
goal is to optimize opcodes/calibrate benchmarks to have the same
execution time.
- Modified the formula to calculate the dependent cost. Instead of the
predicted value, we use the value from the worst scenario. But we also
need to consider that the cost of most opcodes rises logarithmically.
So, there is a point after which each new element is relatively cheap.
And if the cost includes cases with a small number of elements as a
`base` cost, we can make the cost per element much cheaper. So, the
initial benchmarks with `10`, `1000`, and `10000` elements can be
included in the `base` cost to decrease the cost for each new element.
The predicted value from linear regression is used to find *this* point.

---------

Co-authored-by: tirkesi <jocosa344@gmail.com>