Add query and handling for `MessageStatus` (#1371)
The bridge UI requires an client endpoint for querying the status of a
specific `Message`, whether it has been spent or not. An additional
variant is added for unknown messages, i.e. messages that can't be found
in the DB.

The lookup is by `Nonce`. I'm not sure if we would prefer some other
discriminators, but that seems like the obvious choice when you're
searching the entire space--IDK if it's the most efficient though.

https://github.com/FuelLabs/fuel-core/issues/1323

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>