feat: Retrieve finalized blocks from Ethereum (#1399)
Related issues:
- Closes https://github.com/FuelLabs/fuel-core/issues/1336

This PR changes the approach a Fuel node uses to determine the finalized
blocks on the Ethereum blockchain.

**The existing approach:**
- Assume a "finalization period" - a period of time after which we
assume a block is finalized. Generally, we assume a block is finalized
after 100 new blocks have been produced afterwards
- Observe what the current block height is as reported by the Ethereum
client
- Subtract the finalization period from the current height to get an
assumed finalized block, i.e. current block height - 100

The problem with this approach, as outlined by Trail of Bits, is that
finality cannot be guaranteed simply by the amount of time passed alone.

**This PR proposes a more robust approach:**
- Use the Ethereum client to query for the most recently finalized block
directly. This is done using the `Finalized` tag in the query. See
https://github.com/gakonst/ethers-rs/pull/1792

This approach removes any assumption on our side as to what the
finalized block is, and gets the canonically accepted finalized block
from Ethereum.