Moved chain specification into `fuel-core-bin` crate (#1792)
The `0.24.0` release
[failed](https://github.com/FuelLabs/fuel-core/actions/runs/8489604374/job/232600247460)
because chain specification is not part of the `fuel-core-bin` crate.

After the latest changes, the `fuel-core` binary uses testnet
configuration as a default configuration in the case when the snapshot
path is not specified by the user. Because the configuration lives
outside of any crate, it is not possible to load it with
`include_bytes!`. This change moves chain specification folders into a
`fuel-core-bin` crate.

It will affect deployments because now the path is
`bin/fuel-core/chainspec/testnet` instead of
`deployment/scripts/chainspec/testnet`.

cc @rfuelsh @liviudm @tirkesi