Refactor/separate validation from other executions (#1837)
Part of: https://github.com/FuelLabs/fuel-core/issues/1751

`Executor` is use both for producing a block, executing dry-run txs, and
validating blocks. This is done using the same functions and we have
internal conditional logic for each use case. This makes the code more
complicated than it needs to be and makes it hard to read.

This PR removes the `Validation` variant from `ExecutionTypes` and
creates new methods for just validation. Since there is only a set
number of places in our code that uses this variant, it cleans the code
up a lot to just have it be separate method, thus removing the need for
all the internal conditional logic for each variant.

Remaining work to do in subsequent PRs:
- Remove block production code from validation branch (Just use Block
everywhere instead of Partial Block and Components, etc)
- Separate `DryRun` from `Production` and remove `ExecutionType`
- Hopefully remove `ExecutionKind` as well. It goes deep so I still need
to figure out how to untangle it.

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Hannes Karppila <hannes.karppila@gmail.com>