Updated block importer to allow more blocks to be queue (#2010)
Before, we required all services to process the import result before
committing a new result. It was done to reduce the gap between on-chain
and off-chain databases and minimize the damage from the
https://github.com/FuelLabs/fuel-core/issues/1584 problem.

The https://github.com/FuelLabs/fuel-core/issues/1584 is no longer a
problem with https://github.com/FuelLabs/fuel-core/pull/2004 fix.
Because of that, we can allow committing more blocks in parallel while
other services are processing old ones. It improves synchronization
speed because we have a buffer before we wait for other services to
catch up. It is very actual for cases when other services are busy right
now with other work, but soon will be available to process
`ImportResult`.

The default value size of the buffer is `1024`.

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: Hannes Karppila <2204863+Dentosal@users.noreply.github.com>