Add syncronization for Gas Price Database and On Chain Database on startup (#2041)
Closes https://github.com/FuelLabs/fuel-core/issues/2030

In the unlikely situation that the metadata db is behind the on chain
database, we should apply the values from each block to the algorithm
updater to produce reasonable values.

For now, we should be able to use the updater deterministically, but
this might become more difficult as we include the DA values in the
updater.

We also define the behavior for when the gas price db is beyond the
on-chain db. For now, we are assuming that the blocks produced with the
gas prices in the db were either reverted or not properly committed to
the on_chain db, and we just revert the gas price db to the height of
the on chain db. i.e. **the on chain db is the source of truth.**

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [ ] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?