feat: add chain config to Docker images (#2052)
This PR adds the primary branch of the
[chain-configuration](https://github.com/FuelLabs/chain-configuration)
repository to all Docker builds. These builds are published as part of
our CI process.

The goal of bundling this data is to remove a runtime dependency on
GitHub.com from our deployment process. Adding this repo to the image
ensures each image has all of the configuration information required to
run a node.