Remove option on shared algo, give default algo, reorder service startup (#2113)
## Linked Issues/PRs
Closes: https://github.com/FuelLabs/fuel-core/issues/2107

## Description
This code will help ensure that some gas price algorithm is provided to
the consumers. We do this by removing the `Option` and always
initializing with our best value. We also prioritize the initialization
of the gas price service.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself

### After merging, notify other teams

- [ ] [TS SDK](https://github.com/FuelLabs/fuels-rs/)

---------

Co-authored-by: green <xgreenx9999@gmail.com>