Txpool v2 features for futures connections with other modules (#2216)
REVIEWER NOTE: Review https://github.com/FuelLabs/fuel-core/pull/2193
first for context
## Linked Issues/PRs
Closes https://github.com/FuelLabs/fuel-core/issues/2187

## Description

(2k of additions is the tests from TxPool v1 which has been instantiated
again)

This PR adds all the logic for the service and shared state of the
TxPool to be able to be used by all the others modules and to make all
the auxiliary mechanisms working.

The PR changes the way the `SharedState` interact with the pool and the
service. Now the `SharedState` only sends commands to the service that
will answer him whenever he is time to do depending on the priority of
the request (managed through a tokio select biaised). This makes all
methods to return async even `find` which I'm not a big fan of but
didn't found any better solution.
**EDIT: I have made some tests especially tries to run `execute_suite`
tests and it seems that blocking can cause "double blocking forever" and
so I'm up for an alternative**

Here is a list of all interactions made by the different modules and
their new name/behavior/reason :

### Block producer
| Existing | New |
| - | - |
| select_transactions(max_gas) | select_transactions(max_gas) |

### PoA
| Existing | New |
| - | - |
| total_consumable_gas() | Nothing (unused) |
| transaction_status_events() | Replaced by new_txs_notifier
[reason](https://github.com/FuelLabs/fuel-core/blob/d5ba082264e1e3db54c6c81f6036eb06b9d3044c/crates/services/consensus_module/poa/src/service.rs#L590)
|
| pending_number() | Not used anymore after transaction_status_events()
change |
| remove_txs() | broadcast_txs_skipped_reason() |

### Block Importer
| Existing | New |
| - | - |
| Stream with blockEvents | Stream with block events |

### P2P
| Existing | New |
| - | - |
| gossiped_tx_stream | tx_from_p2p_stream |
| new_tx_gossip_subscription | new_peers_subscribed_stream |
| broadcast_transaction | broadcast_transaction |
| get_tx_ids| get_tx_ids |
| get_txs | get_txs |
| notify_gossip_transaction_validity |
notify_gossip_transaction_validity |

### GraphQL API

| Existing | New |
| - | - |
| transaction | transaction |
| tx_update_subscribe | tx_update_subscribe |


### Others changes

- The PR also changes the error type to help separate subdomains of
errors and try to be more clear about the moment errors can happens.
- The TTL management of the transactions is also in place as the less
priority task for the service.
- A new thread pool has been added to manage P2P sync that could take a
long time (same behavior as txpool v1)
- The transactions are now stored as Arc because they need to be
shareable with others modules (API, P2P)
- All the tests from the previous pool has been instantiated and pass. 
- The TxPool don't have anymore unused authorization and match the
correct linting requirements.


## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>