Reject queries that are recursive during the resolution (#2330)
If during resolution of the query we hit the same type, reject further
execution. The code to do this check uses `Visitor` logic from the
`async-graphql` library, but this logic is private, so I duplicated it
in our branch, Later it would be nice to create PR to them in make it
public. Or add a way to pass our own query validators into the
`async-graphql`(which will be more performant).

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: AurelienFT <aurelien.foucault@epitech.eu>
Co-authored-by: Mårten Blankfors <marten@blankfors.se>
Co-authored-by: AurelienFT <32803821+AurelienFT@users.noreply.github.com>
Co-authored-by: AurelienFT <aurefook@gmail.com>