Resolve some falky tests and improve CI times (#2401)
## Linked Issues/PRs
Fix https://github.com/FuelLabs/fuel-core/issues/2408
Fix https://github.com/FuelLabs/fuel-core/issues/2407
Fix https://github.com/FuelLabs/fuel-core/issues/2406
Fix https://github.com/FuelLabs/fuel-core/issues/2351

Fix https://github.com/FuelLabs/fuel-core/issues/2393
Fix https://github.com/FuelLabs/fuel-core/issues/2394
Fix https://github.com/FuelLabs/fuel-core/issues/2395

## Description

This PR fix an issue in P2P heartbeat. The problem was that P2P
heartbeat was updated only if new blocks were received or produced. This
means that if we start the node from an existing db but doesn't produce
blocks and not connect it to anyone it will send block height 0 to the
peers that connects to him. We believe that this fix, resolves #2408
#2407 #2406 and #2351.

For #2394 we just increased the timeouts.
For #2393 we removed the panic in the test and just let p2p reconnect
For #2395 we launch this test using multi-threads mode of Tokio to
follow the convention of all the others tests that launch a node using
`FuelCoreDriver`. Also we added a kill of the driver to try to kill the
node in a more graceful way in all of the test, it should fix a lot of
flakyness in these tests

This PR also change the CI workflow by removing all docker related jobs
and codecov job. These two set of jobs has been moved to separated
workflow that are not triggered automatically but can be triggered
manually on the "Actions" tab of this repository (after the merge of
this PR).

The tests launched by the CI job now use `nextest` that allow us to add
timeout for each test and provide more detailed output. The timeout is
currently 5 min (and 8 for two really big tests) because we have tests
that take a long time but we should lower it in the future.
The steps on the matrix are not cancelled anymore when one failed to
allow possible other success and cache their success for a relaunch of
the tests.

There is still more improve to do on our tests especially on timeout and
rapid execution but this should improve a lot our workflow.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: green <xgreenx9999@gmail.com>