Fix last iteration in sequential opcode (#2479)
Fixes https://github.com/FuelLabs/fuel-core/issues/2022

Previously the read and write of the last key of the contract state was
creating error because of multiple bad checks of limits. Also the key
was increase in all case at the end of the loop causing the `increase()`
to fail on overflow even if there is values to be inserted left.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>