Make the rocksdb cache optional in config and add policy for column opening  (#2526)
## Description
Cache is now an option and can be enable or disabled for each rocksdb
instances.

A new rocksdb config has been added to choose between the mode that open
columns on database in a lazy way (useful for tests) and the mode that
open all columns at once at start (useful for production and benches).
Benches now use the `rocksdb-production` feature of fuel-core.

In order to be more future-proof on the database configuration options,
a `DatabaseConfig` structure has been created and is used across the
code.

## Breaking change

Before, if `max-database-cache-size` was unspecified then
`DEFAULT_DATABASE_CACHE_SIZE` was used.
Now it make the cache disabled.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Aaryamann Challani <43716372+rymnc@users.noreply.github.com>