Pagination queries for `balances` endpoint (#2490)
Closes [#2023](https://github.com/FuelLabs/fuel-core/issues/2032)
Closes https://github.com/FuelLabs/fuel-core/issues/2496

## TODO:
* [X] Add integration test due to custom logic related to base asset id
* [X] Make it work properly with base asset id

## Description
This PR enables the paginated queries for balances, but only if the
balance indexation is enabled. Otherwise, the attempt to send paginated
query will result with `Can not use pagination when balances indexation
is not available` error.

Additionally, if the balances indexation is available the default query
cost for "balances" will be set to 0, and the complexity will be
calculated based on the amount of items requested. When indexation is
not available, the default query cost stays intact (40001).

## Checklist
- [X] New behavior is reflected in tests

### Before requesting review
- [X] I have reviewed the code myself

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>
Co-authored-by: AurelienFT <32803821+AurelienFT@users.noreply.github.com>