Execution tracing: GraphQL query to get storage inputs for past blocks (#2491)
## Description

To support local debugging and execution tracing, we need to add an API
that returns the state that VM needs to execute transactions. This is
done by recording all database accesses during tx validation, called
*storage read replay* (bikeshedding in progress). This data is then
exposed as-is through GraphQL. The API is non-trivial to consume, but a
seperate client is proviced, see
https://github.com/FuelLabs/execution-trace

The implementation exposes database table names and representations
directly. Maintaining backwards compatibility with this could turn out
to be quite hard.

Requires `--historical-execution` flag to enable, as this is otherwise
quite expensive.