Change management of unreleased section of CHANGELOG (#2723)
## Linked Issues/PRs
https://github.com/FuelLabs/fuel-core/discussions/2658

## Description
This PR implements the new way of handling the unreleased section of the
changelog in order to avoid conflicts on each PR merge. Details are
available here : https://github.com/FuelLabs/fuel-core/discussions/2658.

Main changes : 

### Split of all unreleased changes in different files

This PR introduces the `.changes/` folder that contains a sub-folder for
each subsection of the changelog (added, breaking, ...). When you
creates a PR you will need to create a file with the name :
`[pr_number].md` and add the text related to your changes in this file.

The job "check-changelog" has been adapted to check for this file.
You can see 3 example of usage here : 

- https://github.com/OltaLabs/no-conflict-keep-a-changelog-action/pull/7
- https://github.com/OltaLabs/no-conflict-keep-a-changelog-action/pull/9
-
https://github.com/OltaLabs/no-conflict-keep-a-changelog-action/pull/10

The "no changelog" label is still working. Example : 

- https://github.com/OltaLabs/no-conflict-keep-a-changelog-action/pull/8

### New GitHub action to edit the changelog automatically on release

When tagging a PR with a new label "pr release" it will trigger a job
that will remove all changes in the `.changes` folder and create the
section in the CHANGELOG.md file. Example :
-
https://github.com/OltaLabs/no-conflict-keep-a-changelog-action/pull/11

### New PR templates

The GitHub action that creates the section in changelog require to know
the version needed to create. In order to provide that information to
the CI, a new PR template has been added that will be automatically
completed by the Action with all the changelog of this version. Example:
![Screenshot 2025-02-17
154732](https://github.com/user-attachments/assets/2d191f03-49ca-42f6-8a49-19004a3cb2da)

GitHub doesn't have an UI to choose between PR template (such as they
have for issues) and so the new default template PR contains two
clickable links (under the "Preview" section) to create a selection menu
between pull requests templates.

## Side modification

I also edited the name and email of all of our Action that commit to the
repository to be the correct github action bot and being unified

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here