Remove p2pservice networkcodec trait (#2388)
## Linked Issues/PRs
<!-- List of related issues/PRs -->

Closes #2368 

## Description
<!-- List of detailed changes -->

In the P2P service, we couple together GossipSub and RequestResponse
codecs in a `NetworkCodec` trait. In practice, this coupling is not
necessary as the two codecs use different Messages and different logics
for encoding/deconding messages. The only thing that they share in
common is using Postcard as the data format to serialize and deserialize
data, but this is not made clear in the network trait.

This PR is a rework of Codecs in the libp2p service, which deletes the
`NetworkCodec` trait and therefore the coupling between `GossipSubCodec`
and `RequestResponseCodec`. In particular:
- [ ] A new DataFormat trait is defined for serializing deserializing
any SerDe datatype, and the helper functions for postcard serialization
and deserialization` are moved to a struct `Postcard` that implements
the `DataFormat` trait,
- [ ] A new `BoundedCodec` struct that implements the
`RequestResponse::Codec` trait for any `DataFormat` supersedes the old
`PostcardCodec`.
- [ ] A new `UnboundedCodec` struct that implements the `GossipsubCodec
trait for any `DataFormat` supersedes the old `PostcardCodec`,
- [ ] 
- [ ] Parts specific to the Request/Response protocol have been moved to
a new `request_response::protocols` module`,
- [ ] The FuelBehaviour and LibP2P service have been refactored to use
different codecs for `Gossipsub` and `RequestResponse`.

TODO: 
- [ ] Better Documentation
- [x] Better names for modules
- [ ] Maybe move DataFormats to a different part of the codebase?

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [ ] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>