Add new Preconfirmation Task (#2756)
## Linked Issues/PRs
<!-- List of related issues/PRs -->
Part of https://github.com/FuelLabs/fuel-core/issues/2739

## Description
<!-- List of detailed changes -->
Introduces new `PreConfirmationSignatureTask` which will be responsible
for receiving Vecs of txs from the producer, signing the txs, and
broadcasting them to the network. The signatures will be with delegate
keys that will be rotated on a regular frequency.

This does not implement the interfaces for performing the actions,
rather defines the tests for the domain behavior of the service.

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>