add implementations for pre conf task (#2780)
## Linked Issues/PRs
<!-- List of related issues/PRs -->
part of https://github.com/FuelLabs/fuel-core/issues/2739

## Description
<!-- List of detailed changes -->
Introduced some working and some dummy implementations. 

We still need to implement the P2P broadcast trait and the delegate key
generator stuff. We don't know what kind of keys we are going to use, so
that needs to be decided on.

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here
https://github.com/FuelLabs/fuel-core/issues/2782
https://github.com/FuelLabs/fuel-core/issues/2783

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>