GraphQL API changes to support pre-confirmations (#2752)
Closes https://github.com/FuelLabs/fuel-core/issues/2771

## Description
This PR contains the changes to the GraphQL API supporting the upcoming
pre-confirmations feature.

Summary:
1. `TransactionStatus` has grown three new variants:
`SuccessDuringBlockProductionStatus`,
`SqueezedOutDuringBlockProductionStatus` and
`FailureDuringBlockProductionStatus`
1. these new variants should be considered as pre-confirmations, i.e.
they will be gossiped and provided by the client even before the entire
block containing the transaction is produced. Technically, they will be
signed and gossiped as soon as the transaction was executed.
2. the old variants will be created in exactly the same way as they were
before this change
1. Introduction of `TransactionStatusStorage` type, which is just an old
`TransactionStatus` under the new name. Ultimately, we need to convert
it to a new type which only contains `Success` and `Failure` variants,
since only these two states are only ever persisted in the DB

### Before requesting review
- [X] I have reviewed the code myself

### After merging, notify other teams

- [X] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)

---------

Co-authored-by: green <xgreenx9999@gmail.com>