Classify the transactions statuses related to preconfirmations as interim (#2865)
Closes #2852 

## Description
This PR updates the API to treat the various statuses as final, and
other as intermittent. For example, the `transaction_status_change`
subscription will no longer close the stream upon receiving the
`PreconfirmationSuccess`, it'll do it only for the truly final statuses:
`Failed`, `Success`, `SqueezedOut`, `PreconfirmationSqueezedOut`.

Additionally, it removes the`PreconfirmationSqeezedOut` variant from the
user-facing transaction status types.

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [X] I have reviewed the code myself

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>