Listen status updates from in TxStatusManager in TxPool (#2882)
## Linked Issues/PRs
Resolves #2860 

## Description
In this PR the `TxPool Service` is receiving all the status updates from
`TxStatusManager` and only send the useful ones to the `Pool Worker` (to
not over use the worker thread) that can use them to update his internal
state.

I also made some naming refactoring to improve TxPool readability.

I also added an update on status of dependencies of a skipped
transaction

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: Mitch Turner <james.mitchell.turner@gmail.com>
Co-authored-by: Rafał Chabowski <rafal.chabowski@fuel.sh>
Co-authored-by: Green Baneling <XgreenX9999@gmail.com>
Co-authored-by: Rafał Chabowski <88321181+rafal-ch@users.noreply.github.com>