Improve TxStatusManager tests coverage (#2911)
## Linked Issues/PRs
Closes https://github.com/FuelLabs/fuel-core/issues/2838.

THIS PR IS A REVIVAL OF https://github.com/FuelLabs/fuel-core/pull/2888
FROM RAFAL. that we can't merge anymore because of CLA. Kudos to RAFAL !

## Description
This PR adds some more coverage to the TxStatusManager. Details are
available in
https://github.com/FuelLabs/fuel-core/issues/2838#issuecomment-2736390453.

The two outstanding TODOs should be re-evaluated after we are ready with
the ultimate integration test for preconfirmations, because it's likely
that this test will cover such cases.

### Before requesting review
- [x] I have reviewed the code myself


Side note as explained above: this is a copy-paste + some reviews
comments directly addressed from #2888